% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Drwas the timers series of a certain name of a certain sex}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{the name that is going to be plot}

\item{the_sex}{the specified sex for the name being plotted}
}
\value{
the time series of the name and sex
}
\description{
Drwas the timers series of a certain name of a certain sex
}
\examples{
dontrun {draw_a_name(c("Marie", "F"))}
}
