% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Draws the time series of a set of names}
\usage{
draw_names(names)
}
\arguments{
\item{names}{the names that are going to be plot}
}
\value{
the plot of the time series
}
\description{
Draws the time series of a set of names
}
\examples{
dontrun {draw_names(c("Marie", "Pierre"))}
}
