% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{Imports the information of all the shees in an excel file}
\usage{
multi_excel(File)
}
\arguments{
\item{File}{Excel file with all the information}
}
\value{
List of with the information of all the sheets
}
\description{
Imports the information of all the shees in an excel file
}
\examples{
dontrun {multi_excel("Documents/ExcelFiles")}
}
