% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{Imports in a list the informat of all the csv in a folder}
\usage{
my_csv_reader(Folder)
}
\arguments{
\item{Folder}{Folder with the list of csv files}
}
\value{
a list with all the information of the csv
}
\description{
Imports in a list the informat of all the csv in a folder
}
\examples{

}
